Update GQL types to reflect predicted margins as floats
I made the change from int to float for predicted margins in the
Prediction model awhile ago, but neglected to update the GQL
types. Changing how some cumulative metrics were calculated
revealed the discrepancy.

+++ backend/server/graphql/schema.py
@@ -59,7 +59,7 @@ class RoundMetricsType(graphene.ObjectType):
         default_value=0,
         required=True,
     )
-    cumulative_margin_difference = graphene.Int(
+    cumulative_margin_difference = graphene.Float(
         description=(
             "Cumulative difference between predicted margin and actual margin "
             "for the given season."
